refactor: 등수 조건 상수화
