+++ src/constants/Condition.js
@@ -4,6 +4,30 @@ const LOTTO = {
   NUMBER_RANGE_MAX: 45,
 };
 
+const PRIZE = {
+  1: {
+    MATCH: 6,
+    REWARD: 2000000000,
+  },
+  2: {
+    MATCH: 5,
+    REWARD: 30000000,
+  },
+  3: {
+    MATCH: 5,
+    REWARD: 1500000,
+  },
+  4: {
+    MATCH: 4,
+    REWARD: 50000,
+  },
+  5: {
+    MATCH: 3,
+    REWARD: 5000,
+  },
+};
+
 export default {
   LOTTO,
+  PRIZE,
 };
